requirements.txt
asgiref==3.6.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.1.6
django-allauth==0.60.1
django-cors-headers==3.13.0
django-environ==0.9.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==2.2.0
